+++ nesting-solutions-backend/src/main/java/com/nestingsolutions/nestingsolutions/dto/IndividualDto.java
@@ -2,11 +2,9 @@ package com.nestingsolutions.nestingsolutions.dto;
 
 import lombok.Data;
 
-import javax.persistence.Column;
-
 @Data
 public class IndividualDto {
-    private Integer id;
+    private Long id;
     private String firstName;
     private String surname;
     private String email;
