+++ nesting-solutions-backend/src/main/java/com/nestingsolutions/nestingsolutions/dto/RoomTypeDto.java
@@ -2,11 +2,9 @@ package com.nestingsolutions.nestingsolutions.dto;
 
 import lombok.Data;
 
-import javax.persistence.Column;
-
 @Data
 public class RoomTypeDto {
-    private Integer id;
+    private Long id;
     private String category;
     private Integer capacity;
 }
